requirements.txt
Adafruit-Blinka==8.66.1
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-rgb-display==3.14.1
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.83.2
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
fonttools==4.60.0
gpiozero==2.0.1
kiwisolver==1.4.9
lgpio==0.2.2.0
matplotlib==3.10.6
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyftdi==0.57.1
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyusb==1.3.1
six==1.17.0
sysv-ipc==1.1.0
typing_extensions==4.15.0
tzdata==2025.2
